+++ scripts/compile_cost_assumptions.py
@@ -289,7 +289,7 @@ manual_cost_year_assignments_2020 = [
     "gas storage charger",
     "gas storage discharger",
     "gas storage discharger",
-    "perennials gbr",
+    "perennials refining",
 ]
 
 cost_year_2019 = [
@@ -3399,7 +3399,7 @@ def add_biomethanation_CO2(
     return new_technology_dataframe
 
 
-def add_perennials_gbr(
+def add_perennials_refining(
     years: list,
     sheet_names_dict: dict,
     new_technology_dataframe: pd.DataFrame,
@@ -3426,10 +3426,10 @@ def add_perennials_gbr(
     Returns
     -------
     pandas.DataFrame
-        Updated technology data with "perennials gbr".
+        Updated technology data with "perennials refining".
     """
 
-    tech_name = "perennials gbr"
+    tech_name = "perennials refining"
 
     # References (store also in "source" below)
     source_r1 = "https://doi.org/10.1016/B978-0-323-95879-0.50147-8"
@@ -4989,7 +4989,7 @@ if __name__ == "__main__":
     # add biomethanation from pure CO2
     data = add_biomethanation_CO2(years_list, dea_sheet_names, data, data)
     # add perennials and green biorefining
-    data = add_perennials_gbr(years_list, dea_sheet_names, data, data)
+    data = add_perennials_refining(years_list, dea_sheet_names, data, data)
 
     # adjust for inflation
     for x in data.index.get_level_values("technology"):