+++ docs/release_notes.rst
@@ -11,10 +11,12 @@ Release Notes
 Upcoming Release
 ================
 
-.. .. warning:: 
-  
-..   The features listed below are not released yet, but will be part of the next release! 
+.. .. warning::
+
+..   The features listed below are not released yet, but will be part of the next release!
 ..   To use the features already you have to use the ``master`` branch.
+
+* Renamed the "perennials gbr" technology to "perennials refining" for clarity (cost values unchanged).
 
 `v0.14.0 <https://github.com/PyPSA/technology-data/releases/tag/v0.14.0>`__ (13th February 2026)
 ================================================================================================